Clean up ErrorBoundary code
