+++ src/lib/containers/download_list_v2/DownloadListTableV2.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import { useErrorHandler } from 'react-error-boundary'
 import { DownloadListItemResult } from '../../utils/synapseTypes/DownloadListV2/DownloadListItemResult'
-import { toError } from '../ErrorBanner'
+import { toError } from '../../utils/ErrorUtils'
 import * as ReactBootstrap from 'react-bootstrap'
 import { calculateFriendlyFileSize } from '../../utils/functions/calculateFriendlyFileSize'
 import { useGetAvailableFilesToDownloadInfinite } from '../../utils/hooks/SynapseAPI/useGetAvailableFilesToDownload'
