+++ src/lib/containers/entity_finder/details/view/DetailsViewRow.tsx
@@ -21,7 +21,7 @@ import { Hit } from '../../../../utils/synapseTypes/Search'
 import { VersionInfo } from '../../../../utils/synapseTypes/VersionInfo'
 import { EntityBadge } from '../../../EntityBadge'
 import { EntityTypeIcon } from '../../../EntityIcon'
-import { toError } from '../../../ErrorBanner'
+import { toError } from '../../../../utils/ErrorUtils'
 import UserCard from '../../../UserCard'
 import { Checkbox } from '../../../widgets/Checkbox'
 import { BUNDLE_REQUEST_OBJECT } from '../../EntityFinderUtils'
